+++ tests/testthat/test-PROMAX.R
@@ -30,19 +30,19 @@ test_that("output class and dimensions are correct", {
 })
 
 test_that("settings are returned correctly", {
-  expect_named(prom$settings, c("kaiser", "P_type", "precision",
-                                "order_type", "k"))
-  expect_named(prom_psych$settings, c("kaiser", "P_type", "precision",
-                                      "order_type", "k"))
-  expect_named(prom_spss$settings, c("kaiser", "P_type", "precision",
-                                     "order_type", "k"))
-  expect_named(prom_1$settings, c("kaiser", "P_type", "precision",
-                                  "order_type", "k"))
+  expect_named(prom$settings, c("normalize", "P_type", "precision",
+                                "order_type", "varimax_type", "k"))
+  expect_named(prom_psych$settings, c("normalize", "P_type", "precision",
+                                      "order_type", "varimax_type", "k"))
+  expect_named(prom_spss$settings, c("normalize", "P_type", "precision",
+                                     "order_type", "varimax_type", "k"))
+  expect_named(prom_1$settings, c("normalize", "P_type", "precision",
+                                  "order_type", "varimax_type", "k"))
 
-  expect_equal(prom$settings$kaiser, TRUE)
-  expect_equal(prom_psych$settings$kaiser, TRUE)
-  expect_equal(prom_spss$settings$kaiser, TRUE)
-  expect_equal(prom_1$settings$kaiser, TRUE)
+  expect_equal(prom$settings$normalize, TRUE)
+  expect_equal(prom_psych$settings$normalize, TRUE)
+  expect_equal(prom_spss$settings$normalize, TRUE)
+  expect_equal(prom_1$settings$normalize, TRUE)
 
   expect_equal(prom$settings$P_type, "norm")
   expect_equal(prom_psych$settings$P_type, "unnorm")
@@ -59,6 +59,11 @@ test_that("settings are returned correctly", {
   expect_equal(prom_spss$settings$order_type, "ss_factors")
   expect_equal(prom_1$settings$order_type, "eigen")
 
+  expect_equal(prom$settings$varimax_type, "sv")
+  expect_equal(prom_psych$settings$varimax_type, "svd")
+  expect_equal(prom_spss$settings$varimax_type, "kaiser")
+  expect_equal(prom_1$settings$varimax_type, "svd")
+
   expect_equal(prom$settings$k, 4)
   expect_equal(prom_psych$settings$k, 4)
   expect_equal(prom_spss$settings$k, 4)
@@ -70,24 +75,24 @@ test_that("errors etc. are thrown correctly", {
 
   expect_error(.PROMAX(unrot, type = "none"), ' One of "P_type", "precision", "order_type", or "k" was NULL and no valid "type" was specified. Either use one of "EFAtools", "psych", or "SPSS" for type, or specify all other arguments\n')
 
-  expect_warning(.PROMAX(unrot, type = "EFAtools", kaiser = FALSE), " Type and kaiser is specified. kaiser is used with value ' FALSE '. Results may differ from the specified type\n")
+  expect_warning(.PROMAX(unrot, type = "EFAtools", normalize = FALSE), " Type and normalize is specified. normalize is used with value ' FALSE '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "EFAtools", P_type = "norm"), " Type and P_type is specified. P_type is used with value ' norm '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "EFAtools", precision = 1e-9), " Type and precision is specified. precision is used with value ' 1e-09 '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "EFAtools", order_type = "ss_factors"), " Type and order_type is specified. order_type is used with value ' ss_factors '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "EFAtools", k = 2), " Type and k is specified. k is used with value ' 2 '. Results may differ from the specified type\n")
 
-  expect_warning(.PROMAX(unrot, type = "psych", kaiser = FALSE), " Type and kaiser is specified. kaiser is used with value ' FALSE '. Results may differ from the specified type.\n")
+  expect_warning(.PROMAX(unrot, type = "psych", normalize = FALSE), " Type and normalize is specified. normalize is used with value ' FALSE '. Results may differ from the specified type.\n")
   expect_warning(.PROMAX(unrot, type = "psych", P_type = "norm"), " Type and P_type is specified. P_type is used with value ' norm '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "psych", precision = 1e-9), " Type and precision is specified. precision is used with value ' 1e-09 '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "psych", order_type = "ss_factors"), " Type and order_type is specified. order_type is used with value ' ss_factors '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "psych", k = 2), " Type and k is specified. k is used with value ' 2 '. Results may differ from the specified type\n")
 
-  expect_warning(.PROMAX(unrot, type = "SPSS", kaiser = FALSE), " Type and kaiser is specified. kaiser is used with value ' FALSE '. Results may differ from the specified type.\n")
+  expect_warning(.PROMAX(unrot, type = "SPSS", normalize = FALSE), " Type and normalize is specified. normalize is used with value ' FALSE '. Results may differ from the specified type.\n")
   expect_warning(.PROMAX(unrot, type = "SPSS", P_type = "unnorm"), " Type and P_type is specified. P_type is used with value ' unnorm '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "SPSS", precision = 1e-9), " Type and precision is specified. precision is used with value ' 1e-09 '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "SPSS", order_type = "eigen"), " Type and order_type is specified. order_type is used with value ' eigen '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot, type = "SPSS", k = 2), " Type and k is specified. k is used with value ' 2 '. Results may differ from the specified type\n")
-
+  expect_warning(.PROMAX(unrot, type = "SPSS", varimax_type = "svd"), " Type and varimax_type is specified. varimax_type is used with value ' svd '. Results may differ from the specified type\n")
   expect_warning(.PROMAX(unrot_1, type = "EFAtools"), " Cannot rotate single factor. Unrotated loadings returned.\n")
 })