+++ tests/testthat/test-ROTATE_ORTH.R
@@ -46,19 +46,19 @@ test_that("output class and dimensions are correct", {
 })
 
 test_that("settings are returned correctly", {
-  expect_named(equa$settings, c("kaiser", "precision", "order_type"))
-  expect_named(equa_1$settings,c("kaiser", "precision", "order_type"))
-  expect_named(quarti$settings, c("kaiser", "precision", "order_type"))
-  expect_named(bentT$settings, c("kaiser", "precision", "order_type"))
-  expect_named(geoT$settings, c("kaiser", "precision", "order_type"))
-  expect_named(bifacT$settings, c("kaiser", "precision", "order_type"))
+  expect_named(equa$settings, c("normalize", "precision", "order_type"))
+  expect_named(equa_1$settings,c("normalize", "precision", "order_type"))
+  expect_named(quarti$settings, c("normalize", "precision", "order_type"))
+  expect_named(bentT$settings, c("normalize", "precision", "order_type"))
+  expect_named(geoT$settings, c("normalize", "precision", "order_type"))
+  expect_named(bifacT$settings, c("normalize", "precision", "order_type"))
 
-  expect_equal(equa$settings$kaiser, TRUE)
-  expect_equal(equa_1$settings$kaiser, TRUE)
-  expect_equal(quarti$settings$kaiser, TRUE)
-  expect_equal(bentT$settings$kaiser, TRUE)
-  expect_equal(geoT$settings$kaiser, TRUE)
-  expect_equal(bifacT$settings$kaiser, TRUE)
+  expect_equal(equa$settings$normalize, TRUE)
+  expect_equal(equa_1$settings$normalize, TRUE)
+  expect_equal(quarti$settings$normalize, TRUE)
+  expect_equal(bentT$settings$normalize, TRUE)
+  expect_equal(geoT$settings$normalize, TRUE)
+  expect_equal(bifacT$settings$normalize, TRUE)
 
   expect_equal(equa$settings$precision, 1e-5)
   expect_equal(equa_1$settings$precision, 1e-5)
@@ -79,17 +79,17 @@ test_that("errors etc. are thrown correctly", {
   expect_error(.ROTATE_ORTH(unrot, rotation = "equamax", type = "none"), ' "order_type" was NULL and no valid "type" was specified. Either use one of "EFAtools", "psych", or "SPSS" for type, or specify the "order_type" argument\n')
 
   expect_warning(.ROTATE_ORTH(unrot, rotation = "equamax", type = "EFAtools",
-                              kaiser = FALSE), " Type and kaiser is specified. kaiser is used with value ' FALSE '. Results may differ from the specified type\n")
+                              normalize = FALSE), " Type and normalize is specified. normalize is used with value ' FALSE '. Results may differ from the specified type\n")
   expect_warning(.ROTATE_ORTH(unrot, rotation = "equamax", type = "EFAtools",
                               order_type = "ss_factors"), " Type and order_type is specified. order_type is used with value ' ss_factors '. Results may differ from the specified type\n")
 
   expect_warning(.ROTATE_ORTH(unrot, rotation = "equamax", type = "psych",
-                              kaiser = FALSE), " Type and kaiser is specified. kaiser is used with value ' FALSE '. Results may differ from the specified type\n")
+                              normalize = FALSE), " Type and normalize is specified. normalize is used with value ' FALSE '. Results may differ from the specified type\n")
   expect_warning(.ROTATE_ORTH(unrot, rotation = "equamax", type = "psych",
                               order_type = "ss_factors"), " Type and order_type is specified. order_type is used with value ' ss_factors '. Results may differ from the specified type\n")
 
   expect_warning(.ROTATE_ORTH(unrot, rotation = "equamax", type = "SPSS",
-                              kaiser = FALSE), " Type and kaiser is specified. kaiser is used with value ' FALSE '. Results may differ from the specified type\n")
+                              normalize = FALSE), " Type and normalize is specified. normalize is used with value ' FALSE '. Results may differ from the specified type\n")
   expect_warning(.ROTATE_ORTH(unrot, rotation = "equamax", type = "SPSS",
                               order_type = "ss_factors"), " Type and order_type is specified. order_type is used with value ' ss_factors '. Results may differ from the specified type\n")